+++ examples/hummer/playground/hm.config.js
@@ -2,7 +2,7 @@ const path = require('path')
 module.exports = {
   type: 'hummer',
   webpack: {
-    entries: "src/*/index.ts",
+    entries: "src/**/index.ts",
     output: {
       path: path.resolve(__dirname, './dist'),
       filename: "[name].js"
