+++ examples/hummer/playground/src/components/switch/index.ts
@@ -1,8 +1,8 @@
 import { Hummer, View, Text, Switch } from '@hummer/hummer-front'
-import { GroupTitle, GroupContent } from '../common/CommonLayout'
-import { ComponentPage } from '../common/CommonPage'
-import { Color } from '../common/CommonColor'
-import { Style } from '../common/CommonStyle'
+import { GroupTitle, GroupContent } from '../../common/CommonLayout'
+import { ComponentPage } from '../../common/CommonPage'
+import { Color } from '../../common/CommonColor'
+import { Style } from '../../common/CommonStyle'
 
 class RootView extends ComponentPage {
   constructor() {
