+++ examples/hummer/playground/src/components/text/index.ts
@@ -1,10 +1,50 @@
-import { Hummer, View, Text, Input } from '@hummer/hummer-front'
-import { GroupTitle, GroupContent } from '../common/CommonLayout'
-import { ComponentPage } from '../common/CommonPage'
-import { Color } from '../common/CommonColor'
-import { Style } from '../common/CommonStyle'
+import { Hummer, View, Text } from '@hummer/hummer-front'
+import { GroupTitle, GroupContent } from '../../common/CommonLayout'
+import { ComponentPage } from '../../common/CommonPage'
+import { Color } from '../../common/CommonColor'
+import { Style } from '../../common/CommonStyle'
 
-export class InputPage extends ComponentPage {
+
+class RootView extends ComponentPage {
+  constructor() {
+    super();
+    this.setPageTitle('Text')
+  }
+
+  initContentView() {
+    super.initContentView();
+
+    // 通用
+    this.initBackgroundSetting(0);
+    this.initCornerRadiusSetting(0);
+    this.initBorderSetting(0);
+    this.initShadowSetting(0);
+
+    // Text
+    this.initTextAlignSetting(1);
+    this.initFontColorSetting(0);
+    this.initFontSizeSetting(2);
+    this.initFontStyleSetting(0);
+    this.initTextDecorationSetting(0);
+    this.initTextOverflowSetting(0);
+    this.initTextLineClampSetting(0);
+    this.initLetterSpacingSetting(0);
+    this.initlineSpacingSetting(1);
+  }
+
+  createDisplayView(): View {
+    let displayView = new Text();
+    displayView.text = 'This is a Text!\nThis is a Text!\nThis is a Text!\nThis is a Text!';
+    displayView.style = {
+      width: 220,
+      minHeight: 120,
+      backgroundColor: Color.hm_green,
+      textAlign: 'center',
+      fontSize: 22,
+      color: Color.black,
+    }
+    return displayView;
+  }
 
   initTextAlignSetting(selectIndex: number = 1) {
     this.contentView.appendChild(new GroupTitle('文本对齐'));
@@ -42,7 +82,7 @@ export class InputPage extends ComponentPage {
     this.contentView.appendChild(new GroupTitle('文本颜色'));
 
     let layout = new GroupContent();
-    layout.style = { paddingLeft: 18 };
+    layout.style = { paddingLeft: 20 };
     this.contentView.appendChild(layout);
 
     let views = new Array<View>();
@@ -57,74 +97,8 @@ export class InputPage extends ComponentPage {
 
       view.addEventListener('tap', () => {
         this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).style = {
+        (<Text>this.displayView).style = {
           color: textColors[i],
-        }
-      });
-
-      this.selectItemWithBorder(views, selectIndex);
-
-      layout.appendChild(view);
-    });
-  }
-
-  initPlaceholderColorSetting(selectIndex: number = 0) {
-    this.contentView.appendChild(new GroupTitle('占位提示文本颜色'));
-
-    let layout = new GroupContent();
-    layout.style = { paddingLeft: 18 };
-    this.contentView.appendChild(layout);
-
-    let views = new Array<View>();
-    let placeholderColors: string[] = [Color.transparent_grey, Color.hm_yellow, Color.hm_blue];
-    placeholderColors.forEach((color, i) => {
-      let view = new View();
-      view.style = {
-        width: 36,
-        height: 20,
-        marginTop: 8,
-        marginBottom: 8,
-        backgroundColor: color,
-      }
-      views.push(view);
-
-      view.addEventListener('tap', () => {
-        this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).style = {
-          placeholderColor: placeholderColors[i],
-        }
-      });
-
-      this.selectItemWithBorder(views, selectIndex);
-
-      layout.appendChild(view);
-    });
-  }
-
-  initCursorColorSetting(selectIndex: number = 0) {
-    this.contentView.appendChild(new GroupTitle('光标颜色'));
-
-    let layout = new GroupContent();
-    layout.style = { paddingLeft: 18 };
-    this.contentView.appendChild(layout);
-
-    let views = new Array<View>();
-    let cursorColors: string[] = [Color.black, Color.hm_yellow, Color.hm_blue];
-    cursorColors.forEach((color, i) => {
-      let view = new View();
-      view.style = {
-        width: 36,
-        height: 20,
-        marginTop: 8,
-        marginBottom: 8,
-        backgroundColor: color,
-      }
-      views.push(view);
-
-      view.addEventListener('tap', () => {
-        this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).style = {
-          cursorColor: cursorColors[i],
         }
       });
 
@@ -147,12 +121,13 @@ export class InputPage extends ComponentPage {
       view.text = size.toString();
       view.style = {
         ...Style.SelectTextItemStyle,
+        fontSize: size / 1.5,
       }
       views.push(view);
 
       view.addEventListener('tap', () => {
         this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).style = {
+        (<Text>this.displayView).style = {
           fontSize: fontSizes[i],
         }
       });
@@ -163,26 +138,29 @@ export class InputPage extends ComponentPage {
     });
   }
 
-  initMaxLengthSetting(selectIndex: number = 0) {
-    this.contentView.appendChild(new GroupTitle('最大输入长度'));
+  initFontStyleSetting(selectIndex: number = 0) {
+    this.contentView.appendChild(new GroupTitle('文本样式'));
 
     let layout = new GroupContent();
     this.contentView.appendChild(layout);
 
     let views = new Array<View>();
-    let maxLengths: number[] = [0, 3, 6, 9];
-    maxLengths.forEach((length, i) => {
+    let fontStyles: string[] = ['normal', 'bold', 'italic', 'bold & italic'];
+    fontStyles.forEach((style, i) => {
       let view = new Text();
-      view.text = length.toString();
+      view.text = style;
       view.style = {
         ...Style.SelectTextItemStyle,
+        fontWeight: style === 'bold' || style === 'bold & italic' ? 'bold' : 'normal',
+        fontStyle: style === 'italic' || style === 'bold & italic' ? 'italic' : 'normal',
       }
       views.push(view);
 
       view.addEventListener('tap', () => {
         this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).style = {
-          maxLength: maxLengths[i],
+        (<Text>this.displayView).style = {
+          fontWeight: fontStyles[i] === 'bold' || fontStyles[i] === 'bold & italic' ? 'bold' : 'normal',
+          fontStyle: fontStyles[i] === 'italic' || fontStyles[i] === 'bold & italic' ? 'italic' : 'normal',
         }
       });
 
@@ -192,27 +170,28 @@ export class InputPage extends ComponentPage {
     });
   }
 
-  initTypesSetting(selectIndex: number = 0) {
-    this.contentView.appendChild(new GroupTitle('键盘类型'));
+  initTextDecorationSetting(selectIndex: number = 0) {
+    this.contentView.appendChild(new GroupTitle('文本装饰'));
 
     let layout = new GroupContent();
     this.contentView.appendChild(layout);
 
     let views = new Array<View>();
-    let types: string[] = ['default', 'number', 'tel', 'email', 'password'];
-    type Type = 'default' | 'number' | 'tel' | 'email' | 'password';
-    types.forEach((type, i) => {
+    let textDecorations = ['none', 'underline', 'line-through'];
+    type TextDecoration = 'none' | 'underline' | 'line-through';
+    textDecorations.forEach((decoration, i) => {
       let view = new Text();
-      view.text = type.toString();
+      view.text = decoration;
       view.style = {
         ...Style.SelectTextItemStyle,
+        textDecoration: <TextDecoration>decoration,
       }
       views.push(view);
 
       view.addEventListener('tap', () => {
         this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).style = {
-          type: <Type>types[i],
+        (<Text>this.displayView).style = {
+          textDecoration: <TextDecoration>textDecorations[i],
         }
       });
 
@@ -222,18 +201,18 @@ export class InputPage extends ComponentPage {
     });
   }
 
-  initReturnKeyTypesSetting(selectIndex: number = 0) {
-    this.contentView.appendChild(new GroupTitle('键盘返回按钮类型'));
+  initTextOverflowSetting(selectIndex: number = 0) {
+    this.contentView.appendChild(new GroupTitle('文本省略'));
 
     let layout = new GroupContent();
     this.contentView.appendChild(layout);
 
     let views = new Array<View>();
-    let returnKeyTypes: string[] = ['done', 'go', 'next', 'search', 'send'];
-    type ReturnKeyTypes = 'done' | 'go' | 'next' | 'search' | 'send';
-    returnKeyTypes.forEach((type, i) => {
+    let textOverflows = ['ellipsis', 'clip'];
+    type TextOverflow = 'ellipsis' | 'clip';
+    textOverflows.forEach((overflow, i) => {
       let view = new Text();
-      view.text = type.toString();
+      view.text = overflow === 'ellipsis' ? overflow + '...' : overflow;
       view.style = {
         ...Style.SelectTextItemStyle,
       }
@@ -241,8 +220,8 @@ export class InputPage extends ComponentPage {
 
       view.addEventListener('tap', () => {
         this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).style = {
-          returnKeyType: <ReturnKeyTypes>returnKeyTypes[i],
+        (<Text>this.displayView).style = {
+          textOverflow: <TextOverflow>textOverflows[i],
         }
       });
 
@@ -252,17 +231,17 @@ export class InputPage extends ComponentPage {
     });
   }
 
-  initFocusSetting(selectIndex: number = 1) {
-    this.contentView.appendChild(new GroupTitle('焦点'));
+  initTextLineClampSetting(selectIndex: number = 0) {
+    this.contentView.appendChild(new GroupTitle('文本行数'));
 
     let layout = new GroupContent();
     this.contentView.appendChild(layout);
 
     let views = new Array<View>();
-    let focuses: string[] = ['focus', 'unfocus'];
-    focuses.forEach((focus, i) => {
+    let textLineClamps: number[] = [0, 1, 2, 3];
+    textLineClamps.forEach((lineClamp, i) => {
       let view = new Text();
-      view.text = focus;
+      view.text = lineClamp.toString();
       view.style = {
         ...Style.SelectTextItemStyle,
       }
@@ -270,7 +249,9 @@ export class InputPage extends ComponentPage {
 
       view.addEventListener('tap', () => {
         this.selectItemWithBorder(views, i);
-        (<Input>this.displayView).focused = focuses[i] === 'focus';
+        (<Text>this.displayView).style = {
+          textLineClamp: textLineClamps[i],
+        }
       });
 
       this.selectItemWithBorder(views, selectIndex);
@@ -279,8 +260,37 @@ export class InputPage extends ComponentPage {
     });
   }
 
-  initEnableSetting(selectIndex: number = 0) {
-    this.contentView.appendChild(new GroupTitle('可用性'));
+  initLetterSpacingSetting(selectIndex: number = 0) {
+    this.contentView.appendChild(new GroupTitle('字间距'));
+
+    let layout = new GroupContent();
+    this.contentView.appendChild(layout);
+
+    let views = new Array<View>();
+    let letterSpacings: number[] = [0, 0.5, 1];
+    letterSpacings.forEach((spacing, i) => {
+      let view = new Text();
+      view.text = spacing.toString();
+      view.style = {
+        ...Style.SelectTextItemStyle,
+      }
+      views.push(view);
+
+      view.addEventListener('tap', () => {
+        this.selectItemWithBorder(views, i);
+        (<Text>this.displayView).style = {
+          letterSpacing: letterSpacings[i],
+        }
+      });
+
+      this.selectItemWithBorder(views, selectIndex);
+
+      layout.appendChild(view);
+    });
+  }
+
+  initlineSpacingSetting(selectIndex: number = 1) {
+    this.contentView.appendChild(new GroupTitle('行间距'));
 
     let layout = new GroupContent();
     layout.style = {
@@ -289,24 +299,27 @@ export class InputPage extends ComponentPage {
     this.contentView.appendChild(layout);
 
     let views = new Array<View>();
-    let enables: string[] = ['enable', 'disable'];
-    enables.forEach((enable, i) => {
+    let lineSpacings: number[] = [0.5, 1, 1.5, 2];
+    lineSpacings.forEach((spacing, i) => {
       let view = new Text();
-      view.text = enable;
+      view.text = spacing.toString();
       view.style = {
         ...Style.SelectTextItemStyle,
-        color: enable !== 'disable' ? Color.black : Color.light_grey,
       }
       views.push(view);
 
       view.addEventListener('tap', () => {
         this.selectItemWithBorder(views, i);
-        this.displayView.enabled = enables[i] !== 'disable';
+        (<Text>this.displayView).style = {
+          lineSpacingMulti: lineSpacings[i],
+        }
       });
 
-      this.selectItemWithBorder(views, selectIndex);
+      this.selectItemWithBorder(views, 1);
 
       layout.appendChild(view);
     });
   }
 }
+
+Hummer.render(new RootView());
